The Great Refactor
Refactored token enum to have nested structs in tuple variants.

+++ src/interpreter/grammar/literal.rs
@@ -3,7 +3,7 @@ pub enum Literal {
   // Literals
   Identifier { lexeme: String },
   String { lexeme: String },
-  Number { lexeme: f64 },
+  Number { lexeme: i64 },
   Void,
 }
